fix(users): return 404 for not-found errors and finalize account deletion housekeeping
### Functional Changes
- The Ban, Unban, and Admin Delete User endpoints now correctly return `404 Not Found` (instead of `400 Bad Request`) when the target user does not exist.
- The account deletion grace period now supports fractional hour values, allowing finer-grained configuration (e.g. for testing).
- Requesting account deletion now immediately invalidates the cached user profile, so subsequent reads reflect the pending deletion status without waiting for cache expiration.

### Technical Changes
- Injected `ICacheService` into `RequestAccountDeletionCommandHandler` to remove `CacheKeys.User(user.Id)` after a successful deletion request.
- Changed `AccountDeletionOptions.GracePeriodHours` from `int` to `double`.
- Reorganized API request DTOs (`UpdateUserProfileRequest`, `InitiateEmailChangeRequest`, `ReserveBookingRequest`) into dedicated `Requests` subfolders/namespaces for better project structure.
- Updated `User.Create` call sites in `CancellationPolicyEngineTests` and `UpdateUserProfileTests` to pass the newly required `Role.Guest` parameter.
- Updated `FunctionalTestWebAppFactory` to seed test users via the `/register/guest` endpoint instead of the deprecated `/register` endpoint.
- Added `UserBanTests` covering ban/unban status transitions, ban count increments, and domain event publication.

+++ src/Bookify.Application/Options/AccountDeletionOptions.cs
@@ -2,5 +2,5 @@ namespace Bookify.Application.Options;
 
 public sealed class AccountDeletionOptions
 {
-    public int GracePeriodHours { get; set; } = 72;
+    public double GracePeriodHours { get; set; } = 72;
 }
